Revert "merge with upstream (#6)"
This reverts commit 28c0e72d1f9a44706b81411853f23d823c4e7f98.

+++ src/gsm.c
@@ -138,14 +138,14 @@ void PrepareGSM(char *cmdline)
 
     FIELD_fix = gGSMFIELDFix != 0 ? 1 : 0;
 
-    sprintf(cmdline, "%hhu %hhu %hhu %llu %llu %hu %u %u %d %d %d", predef_vmode[gGSMVMode].interlace,
+    sprintf(cmdline, "%hhu %hhu %hhu %llu %llu %hu %d %d %d %d %d", predef_vmode[gGSMVMode].interlace,
             predef_vmode[gGSMVMode].mode,
             predef_vmode[gGSMVMode].ffmd,
             predef_vmode[gGSMVMode].display,
             predef_vmode[gGSMVMode].syncv,
             ((predef_vmode[gGSMVMode].ffmd) << 1) | (predef_vmode[gGSMVMode].interlace),
-            (u32)gGSMXOffset,
-            (u32)gGSMYOffset,
+            gGSMXOffset,
+            gGSMYOffset,
             k576p_fix,
             kGsDxDyOffsetSupported,
             FIELD_fix);
